requirements.txt
polygon-api-client>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.65.0
scipy>=1.11.0
kaleido>=0.2.1
